[FEAT]#78: BuildConfig에 추가해요
