+++ data/build.gradle.kts
@@ -1,3 +1,8 @@
+import org.jetbrains.kotlin.konan.properties.Properties
+
+val properties = Properties()
+properties.load(rootProject.file("local.properties").inputStream())
+
 plugins {
     alias(libs.plugins.wespot.android.library)
     alias(libs.plugins.wespot.android.hilt)
@@ -5,6 +10,15 @@ plugins {
 
 android {
     namespace = "com.bff.wespot.data"
+
+    defaultConfig {
+        buildConfigField("String", "MOCK_BASE_URL", properties.getProperty("MOCK_BASE_URL"))
+        buildConfigField("String", "BASE_URL", properties.getProperty("BASE_URL"))
+    }
+
+    buildFeatures {
+        buildConfig = true
+    }
 }
 
 dependencies {
